rename definemap.test.js to test-define.js
