+++ test/test-define.js
